==> explore_figure/covid_validation/01_covid_h5_tp_Seuratobject.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
library(Matrix)
library(rhdf5)
library(tidyverse)
library(glue)
library(here)
library(hdf5r)
library(Seurat)
```

```{r}
# C141
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339769_C141_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val1 <- CreateSeuratObject(data)
val1
```

```{r}
head(val1)
```



```{r}
# C142
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339770_C142_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val2 <- CreateSeuratObject(data)
val2
```

```{r}
# C143
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339771_C143_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val3 <- CreateSeuratObject(data)
val3
```

```{r}
# C144
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339772_C144_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val4 <- CreateSeuratObject(data)
val4
```

```{r}
# C145
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339773_C145_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val5 <- CreateSeuratObject(data)
val5
```

```{r}
# C146
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4339774_C146_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val6 <- CreateSeuratObject(data)
val6
```

```{r}
# C51
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475048_C51_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val7 <- CreateSeuratObject(data)
val7
```

```{r}
val7@meta.data
```


```{r}
# C52
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475049_C52_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val8 <- CreateSeuratObject(data)
val8
```

```{r}
# C100
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475050_C100_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val9 <- CreateSeuratObject(data)
val9
```

```{r}
# C148
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475051_C148_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val10 <- CreateSeuratObject(data)
val10
```

```{r}
# C149
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475052_C149_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val11 <- CreateSeuratObject(data)
val11
```

```{r}
# C152
data <- Read10X_h5(here('data/GSE145926_validate_covid/GSM4475053_C152_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
val12 <- CreateSeuratObject(data)
val12
```

```{r}
# merge SeuratObject
covid_val <- merge(val1, y = c(val2, val3, val4, val5, val6, val7, val8, val9, val10, val11, val12), add.cell.ids = c('val1', 'val2', 'val3', 'val4', 'val5', 'val6','val7', 'val8', 'val9', 'val10', 'val11', 'val12'), project = 'scRNA_IPF_data')
covid_val
```

```{r}
head(covid_val)
```


```{r}
head(covid_val)
```

```{r}
unique(sapply(X = strsplit(colnames(covid_val), split = "_"), FUN = "[", 1))
```

```{r}
saveRDS(covid_val, file = "COVID_validate_h5_to_SeuratObject_merge.RDS")
```
