==> explore_figure/covid_validation/03_Make_Diagnosis_columns.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
library(Matrix)
library(rhdf5)
library(tidyverse)
library(glue)
library(here)
library(hdf5r)
library(Seurat)
```

```{r}
covid_val <- readRDS(here('explore/RDS/COVID_val_Preprocessing.RDS'))
```

```{r}
head(covid_val)
```
```{r}
covid_val$row <- substr(rownames(covid_val@meta.data), 1, 5)
unique(covid_val$row)
```
control - 7 8 9
moderate - 1 2 4 
severe - 3 5 6 10 11 12
```{r}
# Diagnosis
covid_val@meta.data$Diagnosis <- ifelse(covid_val@meta.data$row %in% c("val7_", "val8_", "val9_"), "Control", "NA")
covid_val@meta.data$Diagnosis <- ifelse(covid_val@meta.data$row %in% c("val1_", "val2_", "val4_"), "COVID moderate", covid_val@meta.data$Diagnosis)
covid_val@meta.data$Diagnosis <- ifelse(covid_val@meta.data$row %in% c("val3_", "val5_", "val6_", "val10", "val11", "val12"), "COVID severe", covid_val@meta.data$Diagnosis)
```

```{r}
# HC - healty control 
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val7_"), "HC1", "NA")
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val8_"), "HC2", covid_val@meta.data$Diagnosis_tag)
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val9_"), "HC3", covid_val@meta.data$Diagnosis_tag)

# M - moderate 
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val1_"), "M1", covid_val@meta.data$Diagnosis_tag)
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val2_"), "M2", covid_val@meta.data$Diagnosis_tag)
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val4_"), "M3", covid_val@meta.data$Diagnosis_tag)

# S - severe 
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val3_"), "S1", covid_val@meta.data$Diagnosis_tag)
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val5_"), "S2", covid_val@meta.data$Diagnosis_tag)
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val6_"), "S3", covid_val@meta.data$Diagnosis_tag)
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val10"), "S4", covid_val@meta.data$Diagnosis_tag)
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val11"), "S5", covid_val@meta.data$Diagnosis_tag)
covid_val@meta.data$Diagnosis_tag <- ifelse(covid_val@meta.data$row %in% c("val12"), "S6", covid_val@meta.data$Diagnosis_tag)

```

```{r}
head(covid_val)
```

```{r}
control <- subset(covid_val, subset = Diagnosis == "Control")
moderate <- subset(covid_val, subset = Diagnosis == "COVID moderate")
severe <- subset(covid_val, subset = Diagnosis == "COVID severe")

DimPlot(covid_val, group.by = "Diagnosis", label = T)
DimPlot(covid_val, group.by = "Diagnosis_tag", label = T)
DimPlot(control, group.by = "Diagnosis", label = T)
DimPlot(moderate, group.by = "Diagnosis", label = T)
DimPlot(severe, group.by = "Diagnosis", label = T)

```

```{r}
saveRDS(covid_val, file = "COVID_val_new_column_Diagnosis.RDS")
```
